buttons and LEDs functional
buttons.py runs as a separate process from webcontrol.  start, stop, pause, shutdown buttons supported on pins 16, 21, 20, and 12 respectively.  LED's are on 5,6, and 19, with a remaining possible button on 26.  the button puts data to the flask socket to communicate with the python code.  The get command asks the server for its current gcode line number (app.data.gcodeIndex) and the  app.gcode.uploadFlag and then sets LED's based on the status of the flag and the gcodeIndex.  If flag is 0 it is either stopped or paused.  If index is not zero then it is paused.  the main.py file has the web-side of the code to receive the requests and respond with the machine status.  This code only works if running on the same machine when the buttons.py process.  There are some commented lines for adding the wiimote pendant in the same process to avoid bogging down the rpi with the wii connection timout.  Routines for start, pause, resume, stop, shutodwn are all included.  Move, is also added, but the most useful is the star/stop option.