Fixed moveline string format issue based on python interpreter version
when using python 3.6 or older, % string formatting is allowed.  3.7 and newer the {}.format() function is used.  When main.py starts, it checks the system version, sets the pythonVersion35 flag, then uses the newerformatting if the flag is false.  data.py has the pythonVersion35 flag, Main.py sets the flat, File/gcodeFile.py checks and uses specific formatting in the moveline function if the flag is tru or false.