z axis limit buttons and gcode sending
added buttons on z page for set min, set max, clear z positions.  they are absolutes and will set based on the current z axis location.
added actions in actions.py for each of the buttons and a query button to ask what the values are
added setminz, setmaxz, clearz, getzlimits in actions.  the B commands B17-B20 are interpreted by the controller to be z axis setpoint settings, clearing, or reading.