Rework the update process
This changes the update process so the Python code itself calls the
unzip utility (instead of going through a script). The new code is also
extracted to a temporary directory which is then moved into place after
the current version has been modev away as a backup. I also added a lot
of logging around this area, to have better visibility on what happens.