Spindle speed control support
for firmware supporting spindle speed control, advanced settings for speed control max and min limits as well as settings for the relay type to be active high or active low with spindle speed control.   Relay functions as before, but an additional pwm pin is designated by the firmware to control the speed.  speed is scaled 0 to max and min is the turn-on speed minimum.