Stripped out All Camera and optical calibration calls
commented, but did not remove optical calibration and custom firmware options.  No need to have a camera routine running on systems that might be resource constrained for an option nobody is using.

gcodeclean is called from folder beside webcontrol.  Since it is OS specific, the user must get the proper compiled version for their OS.  the gcodeclean is called via OS.system and automatically changes the file name.  Added the spinning circles for gcodeclean processing wait time, which can be a significant amount of time.